+++ src/plugins/server.js
@@ -1,5 +1,11 @@
 import express from 'express';
 import ExpressPouchDB from 'express-pouchdb';
+import {
+    filter
+} from 'rxjs/operators/filter';
+import {
+    map
+} from 'rxjs/operators/map';
 
 import PouchDB from '../pouch-db';
 import RxError from '../rx-error';
@@ -10,6 +16,7 @@ Core.plugin(ReplicationPlugin);
 
 const APP_OF_DB = new WeakMap();
 const SERVERS_OF_DB = new WeakMap();
+const SUBSCRIPTIONS_OF_DB = new WeakMap();
 
 
 const normalizeDbName = function(db) {
@@ -24,7 +31,22 @@ const getPrefix = function(db) {
     return splitted.join('/') + '/';
 };
 
-export async function spawnServer({
+/**
+ * tunnel requests so collection-names can be used as paths
+ */
+function tunnelCollectionPath(db, path, app, colName) {
+    db[colName].watchForChanges();
+    app.use(path + '/' + colName, function(req, res, next) {
+        if (req.baseUrl === path + '/' + colName) {
+            const to = normalizeDbName(db) + '-rxdb-0-' + colName;
+            const toFull = req.originalUrl.replace('/db/' + colName, '/db/' + to);
+            req.originalUrl = toFull;
+        }
+        next();
+    });
+};
+
+export function spawnServer({
     path = '/db',
     port = 3000
 }) {
@@ -41,32 +63,28 @@ export async function spawnServer({
     const app = express();
     APP_OF_DB.set(db, app);
 
+
     // tunnel requests so collection-names can be used as paths
     // TODO do this for all collections that get created afterwards
-    Object.keys(db.collections).forEach(colName => {
-        db[colName].watchForChanges();
-        console.log(colName);
-        app.use(path + '/' + colName, function(req, res, next) {
-            console.log('#### one req:');
-            console.dir(req.baseUrl);
-            if (req.baseUrl === '/db/' + colName) {
-                console.log('# tunnel:');
-                const to = normalizeDbName(db) + '-rxdb-0-' + colName;
-                const toFull = req.originalUrl.replace('/db/' + colName, '/db/' + to);
-                req.originalUrl = toFull;
-                console.dir(toFull);
-            }
-            next();
+    Object.keys(db.collections).forEach(colName => tunnelCollectionPath(db, path, app, colName));
+
+    // also tunnel if collection is created afterwards
+    // show error if collection is created afterwards
+    const dbSub = db.$
+        .pipe(
+            filter(ev => ev.data.col === '_collections'),
+            map(ev => ev.data.v)
+        )
+        .subscribe(colName => {
+            const err = RxError.newRxError(
+                'S1', {
+                    collection: colName,
+                    database: db.name
+                }
+            );
+            throw err;
         });
-    });
-
-
-    app.use('*', function(req, res, next) {
-        console.log('#### log:');
-        console.dir(req.baseUrl);
-        next();
-    });
-
+    SUBSCRIPTIONS_OF_DB.set(db, dbSub);
 
     app.use(path, ExpressPouchDB(pseudo));
     const server = app.listen(port);
@@ -83,8 +101,10 @@ export async function spawnServer({
  * runs when the database gets destroyed
  */
 export function onDestroy(db) {
-    if (SERVERS_OF_DB.has(db))
+    if (SERVERS_OF_DB.has(db)) {
         SERVERS_OF_DB.get(db).forEach(server => server.close());
+        SUBSCRIPTIONS_OF_DB.get(db).unsubscribe();
+    }
 }